fix(Interaction): recognize touchpad axis movement on oculus touch
The Controller Events script was not recognizing the thumbstick axis
changes on the Oculus Touch controllers unless the touch sensitive
pad at the top of the thumbstick was being touched.

This is because the Controller Events script would force set the
touchpad axis to not being moved if no touch of the button was
found.

This works fine on the Vive wand and is wanted as the vive wand
touchpad axis can wander and report changes in axis when it's
not being touched. But for the Oculus Touch, this additional check
is counter productive.

To fix the issue, the check to see if the touchpad is static or not
has been abstracted into the Controller SDK so each SDK can handle
how they want to determine if the touchpad axis is static or not.

The private method used for comparing the touchpad axis values has now
also been moved to the Shared Methods script as it is a useful generic
method that is now also used by the Controller SDK.